DarkHotel-Capstone/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
einops==0.8.2
voyageai>=0.3.7
fastapi==0.128.0
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
google-auth==2.45.0
google-genai==1.0.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
oauthlib==3.3.1
# onnxruntime==1.23.2  # Removed: no longer needed (was for sentence-transformers)
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
posthog==5.4.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
qdrant-client>=1.7.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
# safetensors==0.7.0  # Removed: no longer needed (was for transformers)
scikit-learn==1.8.0
scipy==1.16.3
# sentence-transformers==5.2.0  # Removed: replaced by voyageai API
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
# tokenizers==0.22.1  # Removed: no longer needed (was for transformers)
# torch==2.9.1  # Removed: no longer needed (was for sentence-transformers)
tqdm==4.67.1
# transformers==4.57.3  # Removed: no longer needed (was for sentence-transformers)
tree-sitter>=0.23.0
tree-sitter-solidity>=1.2.0
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==14.2
zipp==3.23.0
